changed from jsx to tsx
