+++ frontendV2/src/Router.tsx
@@ -1,9 +1,9 @@
 import { Route, Routes } from 'react-router-dom';
 
-import Login from '@pages/LoginScreen';
-import HomeScreen from '@pages/HomeScreen';
-import FAQScreen from '@pages/FAQScreen';
 import DemoScreen from '@pages/DemoScreen';
+import FAQScreen from '@pages/FAQScreen';
+import HomeScreen from '@pages/HomeScreen';
+import LoginScreen from '@pages/LoginScreen';
 import RegisterScreen from '@pages/RegisterScreen';
 
 export default function Router() {
@@ -11,7 +11,7 @@ export default function Router() {
     <Routes>
       <Route index path="/" element={<HomeScreen />} />
       <Route path="/register" element={<RegisterScreen />} />
-      <Route path="/login" element={<Login />} />
+      <Route path="/login" element={<LoginScreen />} />
       <Route path="/demo" element={<DemoScreen />} />
       <Route path="/faq" element={<FAQScreen />} />
     </Routes>
