--- frontendV2/src/components/Footer.jsx
@@ -1,18 +1,0 @@
-import { Container, Row, Col } from 'react-bootstrap';
-
-const Footer = () => {
-  const currentYear = new Date().getFullYear();
-  return (
-    <footer>
-      <Container>
-        <Row>
-          <Col className="text-center py-3">
-            <p>HouseFinder &copy; {currentYear}</p>
-          </Col>
-        </Row>
-      </Container>
-    </footer>
-  );
-};
-
-export default Footer;
