+++ frontendV2/src/components/Footer.tsx
@@ -1,0 +1,18 @@
+import { Container, Row, Col } from 'react-bootstrap';
+
+function Footer() {
+  const currentYear = new Date().getFullYear();
+  return (
+    <footer>
+      <Container>
+        <Row>
+          <Col className="text-center py-3">
+            <p>HouseFinder &copy; {currentYear}</p>
+          </Col>
+        </Row>
+      </Container>
+    </footer>
+  );
+}
+
+export default Footer;
