+++ frontendV2/src/components/FormContainer.tsx
@@ -1,0 +1,15 @@
+import { Container, Row, Col } from 'react-bootstrap';
+
+function FormContainer({ children }) {
+  return (
+    <Container>
+      <Row className="justify-content-md-center">
+        <Col xs={12} md={6}>
+          {children}
+        </Col>
+      </Row>
+    </Container>
+  );
+}
+
+export default FormContainer;
