--- frontendV2/src/constants.js
@@ -1,3 +1,0 @@
-export const BASE_URL = '';
-
-export const USERS_URL = '/api/users';
