+++ frontendV2/src/constants.ts
@@ -1,0 +1,3 @@
+export const BASE_URL = '';
+
+export const USERS_URL = '/api/users';
