--- frontendV2/src/features/apiSlice.js
@@ -1,11 +1,0 @@
-import { createApi, fetchBaseQuery } from '@reduxjs/toolkit/query/react';
-
-import { BASE_URL } from '../constants';
-
-const baseQuery = fetchBaseQuery({ baseURL: BASE_URL });
-
-export const apiSlice = createApi({
-  baseQuery,
-  tagTypes: ['User'],
-  endpoints: (builder) => ({}),
-});
