+++ frontendV2/src/features/apiSlice.ts
@@ -1,0 +1,11 @@
+import { createApi, fetchBaseQuery } from '@reduxjs/toolkit/query/react';
+
+import { BASE_URL } from '../constants';
+
+const baseQuery = fetchBaseQuery({ baseURL: BASE_URL });
+
+export const apiSlice = createApi({
+  baseQuery,
+  tagTypes: ['User'],
+  endpoints: (builder) => ({}),
+});
