--- frontendV2/src/features/usersApiSlice.js
@@ -1,16 +1,0 @@
-import { apiSlice } from './apiSlice';
-import { USERS_URL } from '../constants';
-
-const usersApiSlice = apiSlice.injectEndpoints({
-  endpoints: (builder) => ({
-    login: builder.mutation({
-      query: (data) => ({
-        url: `${USERS_URL}/login`,
-        method: 'POST',
-        body: data,
-      }),
-    }),
-  }),
-});
-
-export const { useLoginMutation } = usersApiSlice;
