+++ frontendV2/src/pages/FAQScreen.tsx
@@ -1,0 +1,11 @@
+function FAQ() {
+  return (
+    <div className="container-xl mx-auto text-center pt-5">
+      <h1 className="text-center pt-4">FAQ</h1>
+      <p>Questions:...</p>
+      <p>Answers:...</p>
+    </div>
+  );
+}
+
+export default FAQ;
