--- frontendV2/src/pages/HomeScreen.jsx
@@ -1,13 +1,0 @@
-import HomeCarousel from '../components/HomeCarousel';
-import About from '../components/About';
-
-function HomeScreen() {
-  return (
-    <>
-      <HomeCarousel />
-      <About />
-    </>
-  );
-}
-
-export default HomeScreen;
