+++ frontendV2/src/pages/HomeScreen.tsx
@@ -1,0 +1,13 @@
+import HomeCarousel from '../components/HomeCarousel';
+import About from '../components/About';
+
+function HomeScreen() {
+  return (
+    <>
+      <HomeCarousel />
+      <About />
+    </>
+  );
+}
+
+export default HomeScreen;
