--- frontendV2/src/pages/RegisterScreen.jsx
@@ -1,65 +1,0 @@
-import { useState } from 'react';
-import { Link } from 'react-router-dom';
-import { Button, Form } from 'react-bootstrap';
-import FormContainer from '../components/FormContainer';
-
-function RegisterScreen() {
-  const [name, setName] = useState('');
-  const [email, setEmail] = useState('');
-  const [password, setPassword] = useState('');
-  const [confirmPassword, setConfirmPassword] = useState('');
-
-  const submitHandler = (e) => {
-    e.preventDefault();
-    console.log('submit');
-  };
-
-  return (
-    <FormContainer>
-      <h3>Sign Up As New Customer</h3>
-      <Form onSubmit={submitHandler}>
-        <Form.Group controlId="name" className="my-3">
-          <Form.Control
-            type="name"
-            placeholder="Enter Name"
-            value={name}
-            onChange={(e) => setName(e.target.value)}
-          />
-        </Form.Group>
-
-        <Form.Group controlId="email" className="my-3">
-          <Form.Control
-            type="email"
-            placeholder="Enter email"
-            value={email}
-            onChange={(e) => setEmail(e.target.value)}
-          />
-        </Form.Group>
-
-        <Form.Group controlId="password" className="my-3">
-          <Form.Control
-            type="password"
-            placeholder="Enter password"
-            value={password}
-            onChange={(e) => setPassword(e.target.value)}
-          />
-        </Form.Group>
-
-        <Form.Group controlId="confirmpassword" className="my-3">
-          <Form.Control
-            type="password"
-            placeholder="Enter password again"
-            value={password}
-            onChange={(e) => setConfirmPassword(e.target.value)}
-          />
-        </Form.Group>
-
-        <Button type="submit" variant="primary" className="mt-3">
-          Sign Up
-        </Button>
-      </Form>
-    </FormContainer>
-  );
-}
-
-export default RegisterScreen;
